% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{backward.selection}
\alias{backward.selection}
\title{Backward selection function for the second screening step}
\usage{
backward.selection(data, pts, lambda, mu, alpha_L = 0.25)
}
\arguments{
\item{data}{A n by p dataset matrix}

\item{pts}{A numeric vector, which includes all candidate change points obtained from the first step}

\item{lambda}{Tuning parameter for sparse component}

\item{mu}{Tuning parameter for low-rank component}

\item{alpha_L}{Constraint space for low rank component, default is 0.25}
}
\value{
A list object, containing
\describe{
    \item{L.n}{Value of objective function}
    \item{L.n.current}{Current value of objective function}
}
}
\description{
Backward elimination algorithm
}
\keyword{internal}
