% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{break.var.lps}
\alias{break.var.lps}
\title{Auxiliary function to calculate loss at the estimated change points}
\usage{
break.var.lps(data, pts, lambda, mu, alpha_L = 0.25)
}
\arguments{
\item{data}{A n by p dataset, n is the number of observations, p is the number of variables}

\item{pts}{a vector of change points}

\item{lambda}{tuning parameter for sparse component}

\item{mu}{tuning parameter for low-rank component}

\item{alpha_L}{a positive number, indicating the value of constraint space of low rank component, default is 0.25}
}
\value{
A list object, including
\describe{
 \item{L.n}{Value of objective function}
}
}
\description{
Function to calculate loss function at the estimated change points
}
\keyword{internal}
