% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{cv.tuning.selection}
\alias{cv.tuning.selection}
\title{a function to apply cross-validation to select tuning parameter by minimizing SSE}
\usage{
cv.tuning.selection(data, lambda.seq, mu.seq, alpha_L = 0.25, nfold = 5)
}
\arguments{
\item{data}{a n by p dataset matrix}

\item{lambda.seq}{a numeric vector, indicates the sequence of tuning parameters of sparse components}

\item{mu.seq}{a numeric vector, the sequence of tuning parameters of low rank components}

\item{alpha_L}{a positive numeric value, indicating the constraint space of low rank components}

\item{nfold}{a positive integer, the number of folds for cv}
}
\value{
a list of object, including
\describe{
    \item{grid}{the grid of lamdbas and mus}
    \item{lambda}{final selected tuning parameter for sparse}
    \item{mu}{final selected tuning parameter for low rank}
}
}
\description{
a function to apply cross-validation to select tuning parameter by minimizing SSE
}
\keyword{internal}
