% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{f.func}
\alias{f.func}
\title{Main loss function for quardratic loss}
\usage{
f.func(x, A, b)
}
\arguments{
\item{x}{Model parameters}

\item{A}{Design matrix with size of n by p}

\item{b}{Correspond vector or matrix}
}
\value{
Value of objective function
}
\description{
Main loss function
}
\keyword{internal}
