% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{gradf.func}
\alias{gradf.func}
\title{Gradient function of quardratic loss}
\usage{
gradf.func(x, AtA, Atb)
}
\arguments{
\item{x}{A vector, or matrix, indicating the model parameter}

\item{AtA}{A p by p Gram matrix for corresponding design matrix A}

\item{Atb}{An inner product for design matrix A and corresponding matrix (vector) b}
}
\value{
Value of gradients
}
\description{
Gradient function of quardratic loss
}
\keyword{internal}
