% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{nuclear.pen}
\alias{nuclear.pen}
\title{Nuclear norm penalty for low-rank component}
\usage{
nuclear.pen(x, lambda)
}
\arguments{
\item{x}{Model parameter}

\item{lambda}{Tuning parameter}
}
\value{
Value of nuclear norm penalty term
}
\description{
Nuclear norm penalty for low-rank component
}
\keyword{internal}
