% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{obj.func}
\alias{obj.func}
\title{Objective function}
\usage{
obj.func(x.lr, x.sparse, A, b, lambda, mu)
}
\arguments{
\item{x.lr}{low-rank component}

\item{x.sparse}{sparse component}

\item{A}{design matrix}

\item{b}{correspond vector}

\item{lambda}{a tuning parameter for sparse component}

\item{mu}{a tuning parameter for low-rank component}
}
\value{
value of objective function
}
\description{
objective function, main loss function and penalties
}
\keyword{internal}
