% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_BSS_SGL.R
\name{pred.block}
\alias{pred.block}
\title{Prediction function (block)}
\usage{
pred.block(Y, phi, q, nob, p, h)
}
\arguments{
\item{Y}{data for prediction}

\item{phi}{parameter matrix}

\item{q}{lag}

\item{nob}{total length of data}

\item{p}{the dimension of time series components}

\item{h}{the length of observation to predict}
}
\value{
prediction matrix
}
\description{
Prediction function (block)
}
\keyword{internal}
