% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_BSS_SGL.R
\name{prox.nuclear.func.fLS}
\alias{prox.nuclear.func.fLS}
\title{Proximal function for nuclear norm penalty}
\usage{
prox.nuclear.func.fLS(y, A, b, L, lambda, AtA, Atb)
}
\arguments{
\item{y}{model parameter}

\item{A}{design matrix}

\item{b}{correspond vector, or matrix}

\item{L}{learning rate}

\item{lambda}{tuning parameter}

\item{AtA}{Gram matrix obtained by design matrix}

\item{Atb}{inner product for design matrix A and correspond vector b}
}
\value{
value of proximal function
}
\description{
Proximal function for nuclear norm penalty
}
\keyword{internal}
