% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{prox.sparse.func}
\alias{prox.sparse.func}
\title{Proximal function with l1-norm penalty updating}
\usage{
prox.sparse.func(w1, y, A, b, L, lambda, AtA, Atb)
}
\arguments{
\item{w1}{previously updated model parameter}

\item{y}{updated model parameter}

\item{A}{design matrix}

\item{b}{correspond vector, or matrix}

\item{L}{learning rate}

\item{lambda}{tuning parameter for sparse component}

\item{AtA}{Gram matrix of design matrix A}

\item{Atb}{inner product of design matrix A and correspond vector b}
}
\value{
Value of proximal function with l1-norm penalty
}
\description{
Proximal function with l1-norm
}
\keyword{internal}
