% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{Q.func}
\alias{Q.func}
\title{An auxiliary function in FISTA algorithm}
\usage{
Q.func(x, y, A, b, L, AtA, Atb)
}
\arguments{
\item{x}{Model parameter for previous update}

\item{y}{Model parameter for updating}

\item{A}{An n by p design matrix}

\item{b}{A correspond vector, or matrix with size of n by 1 or n by p}

\item{L}{Learning rate}

\item{AtA}{Gram matrix for design matrix A}

\item{Atb}{Inner product for design matrix A and correspond vector b}
}
\value{
Value of function Q
}
\description{
Auxiliary function for FISTA implementation
}
\keyword{internal}
