% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{shrinkage.lr}
\alias{shrinkage.lr}
\title{Shrinkage function for low-rank soft-thresholding}
\usage{
shrinkage.lr(y, tau)
}
\arguments{
\item{y}{A matrix, or a vector for thresholding}

\item{tau}{A positive number, threshold}
}
\value{
A thresholded matrix, or vector
}
\description{
Shrinkage function for low-rank soft-thresholding
}
\keyword{internal}
