% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_LpS.R
\name{sparse.pen}
\alias{sparse.pen}
\title{L1-norm penalty for sparse component}
\usage{
sparse.pen(x, lambda)
}
\arguments{
\item{x}{Model parameter}

\item{lambda}{Tuning parameter}
}
\value{
Value of l1-norm penalty term
}
\description{
L1-norm penalty for sparse component
}
\keyword{internal}
