% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions_BSS_SGL.R
\name{third.step.exhaustive.search}
\alias{third.step.exhaustive.search}
\title{Exhaustive search step (third step).}
\usage{
third.step.exhaustive.search(
  data,
  q,
  max.iteration = 1000,
  tol = tol,
  pts.list,
  an,
  phi.est.full = NULL,
  phi.local.1 = NULL,
  phi.local.2 = NULL,
  blocks = NULL
)
}
\arguments{
\item{data}{input data matrix, with each column representing the time series component}

\item{q}{the AR order}

\item{max.iteration}{max number of iteration for the fused lasso}

\item{tol}{tolerance for the fused lasso}

\item{pts.list}{the selected break points clustered by a_n after the second step}

\item{an}{the neighborhood size a_n}

\item{phi.est.full}{list of local parameter estimator}

\item{phi.local.1}{a list of loca parameter estimator}

\item{phi.local.2}{a list of loca parameter estimator}

\item{blocks}{the blocks}
}
\value{
A list object, which contains the followings
\describe{
  \item{pts}{a set of final selected break point after the third exhaustive search step}
}
}
\description{
Perform the exhaustive search to select the break point for each cluster.
}
\keyword{internal}
