% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{weekly}
\alias{weekly}
\title{weekly stock price data}
\format{
An dataframe of weekly stock price data
}
\usage{
data(weekly)
}
\description{
weekly stock price data
}
\examples{
data(weekly)
head(weekly)
}
\keyword{datasets}
